+++ common-lib/esm/t4t-validate.js
@@ -19,23 +19,26 @@
 // - example-vite/src/components/CrudTable.vue
 function validate (rules, type, col, record) {
   let invalid = ''
-  for (let rule in rules) {
-    if (type === 'string') {
-      if (rule === 'min' && record[col].length < rules[rule]) invalid = `need at least ${rules[rule]} characters`
-      else if (rule === 'max' && record[col].length > rules[rule]) invalid = `maximum ${rules[rule]} characters`
-      else if (rule === 'regex' && !RegExp(rules[rule]).test(record[col])) invalid = `regex ${rules[rule]} failed`
-    } else if (['integer', 'decimal', 'datetime', 'date', 'time'].includes(type)) {
-      if (rule === 'min' && record[col] < rules[rule]) invalid = `cannot be less than ${rules[rule]}`
-      else if (rule === 'max' && record[col] > rules[rule]) invalid = `cannot be more than ${rules[rule]}`
-
-      else if (rule === 'gt' && record[rules[rule]] && !(record[col] > record[rules[rule]])) invalid = `${col} must be > ${rules[rule]}`
-      else if (rule === 'gte' && record[rules[rule]] && !(record[col] >= record[rules[rule]])) invalid = `${col} must be >= ${rules[rule]}`
-      else if (rule === 'lt' && record[rules[rule]] && !(record[col] < record[rules[rule]])) invalid = `${col} must be < ${rules[rule]}`
-      else if (rule === 'lte' && record[rules[rule]] && !(record[col] <= record[rules[rule]])) invalid = `${col} must be <= ${rules[rule]}`
-    } else {
-      // nothing here...
-    }
-    if (invalid) break
+  try {
+    for (let rule in rules) {
+      if (type === 'string') {
+        if (rule === 'min' && record[col].length < rules[rule]) invalid = `need at least ${rules[rule]} characters`
+        else if (rule === 'max' && record[col].length > rules[rule]) invalid = `maximum ${rules[rule]} characters`
+        else if (rule === 'regex' && !RegExp(rules[rule]).test(record[col])) invalid = `regex ${rules[rule]} failed`
+      } else if (['integer', 'decimal', 'datetime', 'date', 'time'].includes(type)) {
+        if (rule === 'min' && record[col] < rules[rule]) invalid = `cannot be less than ${rules[rule]}`
+        else if (rule === 'max' && record[col] > rules[rule]) invalid = `cannot be more than ${rules[rule]}`
+        else if (rule === 'gt' && record[rules[rule]] && !(record[col] > record[rules[rule]])) invalid = `${col} must be > ${rules[rule]}`
+        else if (rule === 'gte' && record[rules[rule]] && !(record[col] >= record[rules[rule]])) invalid = `${col} must be >= ${rules[rule]}`
+        else if (rule === 'lt' && record[rules[rule]] && !(record[col] < record[rules[rule]])) invalid = `${col} must be < ${rules[rule]}`
+        else if (rule === 'lte' && record[rules[rule]] && !(record[col] <= record[rules[rule]])) invalid = `${col} must be <= ${rules[rule]}`
+      } else {
+        // nothing here...
+      }
+      if (invalid) break
+    }  
+  } catch (e) {
+    console.log('FE validate', col, e.toString())
   }
   return invalid
 }